put new transport code in modules; write separate examples
